Combine memory pool and query based spill strategies
Always have memory-pool spilling running whenever per-query spilling is
enabled to prevent a situation, where the memory pool could fill up
without spill being triggered. This is a problem because the OOM killer
won't kick in if there is any revocable memory allocated.

The problem exists for the PER_TASK_MEMORY_THRESHOLD spill strategy as
well, but we don't address it here as the fix is more complicated, and
we expect to remove that strategy in the future.