Refactor StreamReader::startStripe to take the Stripe directly
This PR is in preparation for dictionary sharing implementation in
DWRF format, a file format level feature that allows
multiple keys in a flat map to share the same dictionary data stream.
This PR is refactor only and contains no logic change.

Test plan - existing unit tests. We sampled the top 3 namespaces for
tables to run integration tests on and the result is clean.