Avoid ambiguous null value encoding in HivePartitionKey
When encoded as partition names, null column values in Hive use a
magic string: __HIVE_DEFAULT_PARTITION__. In other places though,
null values are represented with the sequence \N. This creates an
unfortunate ambiguity whereby a non-null partition value of "\N"
could subsequently be interpreted as null in some value contexts
such as prefilled values or for the purposes of dynamic filtering.

Instead of converting "__HIVE_DEFAULT_PARTITION__" to "\N" in its
constructor, HivePartitionKey will instead carry Optional<String>
for its value and rely on callers to use the appropriate null
string for their context.