Readjust concurrency and thread pool settings for Presto on Spark tests
The tests finish in less then 20m locally on a machine with nvme flash.
It looks like on the CI the bottleneck is IOPS.

Decreasing task concurrency should result in less granular spills
decreasing overall IO pressure.

Increasing number of threads for spiller is needed to avoid thread
starvation when the spilling is slow.