Change RowExpressionFormatter output to be closer to valid SQL
RowExpressionFormatter outputs the type before any constant
(e.g BIGINT 5) for debugging purposes. This is not valid SQL,
so the constant is now output with TypeConstructor (e.g BIGINT'5'),
except for VARBINARY constants, which use the binary literal
syntax (e.g. X'abc123').

RowExpressionFormatter outputs LIKE expressions such as
  column LIKE 'pattern%'
as
  LIKE(column, CAST(VARCHAR pattern% as LikeExpression))
and LIKE expressions with ESCAPE, such as
  column LIKE '$_pattern%' ESCAPE '$'
as
  LIKE(column, LIKE_PATTERN(VARCHAR pattern%, VARCHAR $))
The output for like expressions was changed to be valid SQL.