Fix broadcast table memory accounting bug in PrestoSparkRemoteSourceOperator
BroadcastTableCacheManager caches HT for a stage so that it can be reused. Memory
used by the cache is accounted for inside PrestoSparkRemoteSourceOperator systemMemoryContext.
The accounting works fine where there is only 1 BHJ in a stage. However, if a stage has N
BHJ, then cache memory is accounted N number of times, which can trigger broadcast OOM.
This PR fixes the accounting bug so that we account for actual memory used by all the
hashtables in the cache properly.