Add task-level RuntimeStats
Add task-level RuntimeStats to track the metrics which are exposed in
Operators.

Add two task-level metrics splitCountPerTask and taskElapsedTimeNanos.
These metrics tracks the sum, min, max and count values and can be used
to detect the skew issues among workers.

Change the metric aggregation to not aggregate metrics across stages as
It doesn't make sense to aggregate metrics of different plan fragments.