Load Hashtables in MAP_AGG
Broadcast join queries with filter regresses when using lazy
hash tables in MapBlockBuilder. Filter repeatedly computes
the hash tables instead of computing it once. This PR forces
the hash tables to be loaded in MAP_AGG to prevent the
regression. This PR is in preparation of making the hash
tables lazy in the MapBlockBuilder.