Introduce commit API for Spiller
Currently, Spillers provide two APIs:
1. spill() which is write API
2. getSpills() which is read API
The write API opens a spill file but does not close it once write is finished.
It is closed when read is triggred. Since, we do not commit the spill file,
sometimes pages are retained in memory buffers which accumulates together if spilling
is triggered too frequently. In low memory situation, this can trigger an OOM
This PR introduces a commit() API in Spiller interface which can be called
after write is done and before read is invoked. This API will flush the
buffers to spill file and close the stream so that resources can be released.