Fix unsafe cast in RowExpressionFormatter
The pattern argument of a LIKE CallExpression
is changed from a CallExpression describing the
pattern into a joni.Regex object when the RowExpression
is optimized by RowExpressionInterperter. When
this occured on a LIKE CallExpression in a query
plan before the plan was output with PlanPrinter,
the following cast error occured:

class com.facebook.presto.spi.relation.ConstantExpression cannot be cast to class com.facebook.presto.spi.relation.CallExpression

This occurred due to an unchecked cast in
RowExpressionFormatter. A type check was added before
the cast, and tests were added to RowExpressionFormatter
that reproduce the behaviour of printing an
optimized LIKE CallExpression.