Disable metadata query rewrite for incomplete stats
Partition Stats stored in metastore may be incomplete or missing, if
stats collection timeout. So even if the metastore has a stats
indicating that the partition is empty, it may not be empty on disk.
This commit disables the metadata query optimization for this case as it
could change the behavior.
To be safe, we only do the rewrite if the stats indicates that the
partition is non-empty. This is the normal case and should cover most
queries in practice.

This check is done separately for the two code paths:
When the query is reducible, we only check the result partition.
When the query is not reducible, we have to check all involved
partitions.