Fix partial aggregation (#10)
Summary:
HashAggregation operator could get into an invalid state where it is not
blocked, doesn't need input, but getOutput() returns null. This happened when
operator processed a "distinct" aggregation and reached a limit on the amount
of memory to use for partial aggregation. This state confused the Driver loop
and caused it to complete prematurely. This would result in either wrong
results or a crash.

The fix is to flush partial aggregation state when reached the limit.

Pull Request resolved: https://github.com/facebookincubator/velox/pull/10

Reviewed By: pedroerp

Differential Revision: D30203061

Pulled By: mbasmanova

fbshipit-source-id: 12a381717462e77d2254490ca01c98920a9b2ad7