Implement Spark's variadic HASH() function (#16)
Summary:
Pull Request resolved: https://github.com/facebookincubator/velox/pull/16

Pull Request resolved: https://github.com/facebookexternal/f4d/pull/1287

Spark uses a variant of Murmur3; I've copied the hash implementation directly
from the Spark codebase, modified of course for C++.

The Spark algorithm for combining hashes is simple: initialize the hash to 42,
and then use the current hash as the seed for the next hash. NULLs are skipped.
This implementation processes one column at a time.

Reviewed By: miaoever

Differential Revision: D29743513

fbshipit-source-id: 0c43b65c6351581e20a84665a77303d55cf61282