Fix HashProbe to load LazyVector before wrapping (#44)
Summary:
HashProbe operator produces results in batches of kOutputBatchSize(=1000) rows
or less. The result of probing an input vector may not fit into a single
batch in which case it is split into multiple batches. For each batch, input
vector is wrapped in a dictionary that repeats a subset of the rows as many
times as there are hits in the hash table.

If input is a lazy vector that hasn't been loaded yet, the join produces two
dictionary vectors wrapping the same lazy vector. When first dictionary
vector is accessed (e.g. in expression evaluation) the underlying lazy vector
is loaded only for a set of rows mentioned in the dictionary. When second
dictionary vector is accessed the underlying lazy vector is already marked as
loaded and no further loading is happening. However, the rows that didn't
make it into the first batch were never loaded and thus are missing.
Accessing these rows causes a crash.

The fix is to not wrap a single lazy vector twice.

Pull Request resolved: https://github.com/facebookincubator/velox/pull/44

Reviewed By: kgpai

Differential Revision: D30342208

Pulled By: mbasmanova

fbshipit-source-id: 2928f08858f40fc03a50f8c2911c953345abde73