Simplified eval path - first cut (#52)
Summary:
Pull Request resolved: https://github.com/facebookincubator/velox/pull/52

Pull Request resolved: https://github.com/facebookexternal/f4d/pull/1353

First cut for a simplified expression evaluation path, which flattens
all vector before execution. The idea is to use this alternative path for tests
and to help isolate production issues. It will also help us understand how much
the current (complex) eval path buys us in terms of performance.
.
A few special form expression are not supported yet, in order to make this diff
a bit smaller.

Reviewed By: mbasmanova, funrollloops

Differential Revision: D30122151

fbshipit-source-id: cdb79ab93a8212ece23c0b07a5e89f7b64403557