Fix ensureWritable() bug affecting expression correctness (#78)
Summary:
Pull Request resolved: https://github.com/facebookincubator/velox/pull/78

Another one found by expression explorer: ensureWritable() considers the
target vector size for constantVectors as being rows.size() (since constantVector
sometimes sets size to kMaxElements). The issue is that ensureWritable() is
often used with SelectivityVector::empty(), which ends up returning a flatten
vector with size() == 0, which breaks the eval.
This diff fixes the errors in the eval, but the long term fix is to prevent
constantVectors from using kMaxElements (and requiring them to use the actual
size).

Reviewed By: mbasmanova

Differential Revision: D30414318

fbshipit-source-id: 6f8660788234bccfdc7eb8a1fc9700a5e626cc87