Rename ChecksumAlgorithm NULL option to NULL_ (#60)
Summary:
Pull Request resolved: https://github.com/facebookincubator/velox/pull/60

gcc on linux (9.3 on Ubuntu 20.04) seems to have a problem with the
`NULL` option name and fails the compilation. Renaming the flag to NULL_
instead. Interestingly enough, all other usages of the flag I could find
already use "NULL_", so I'm a little confused about how this worked in the
first place:

Reviewed By: zzhao0

Differential Revision: D30396818

fbshipit-source-id: 7def0b77955f76fd45af6235a802cf60ac1240e7