Add commented expr eval example binary (#93)
Summary:
Pull Request resolved: https://github.com/facebookincubator/velox/pull/93

Adding a binary containing a commented minimal example of the expr
evaluation engine. The main idea is to provide a similar binary for other
integration models and extensibility model. Hopefully this can make it
easier for Velox users to understand the different usage modes, help
highlight the external APIs, and serve a reference for developers
integrating with Velox.

Reviewed By: kgpai, mbasmanova

Differential Revision: D30452468

fbshipit-source-id: ecb9b37f9825114c8f8e6d524167d72f0620ca63