Implement INSTR for Spark (#73)
Summary:
Pull Request resolved: https://github.com/facebookincubator/velox/pull/73

Presto's `strpos` was originally mapped, but the name of the corresponding
function in Spark is `instr` and it returns an `int32`, not an `int64`. Instead
of moving Presto's implementation to lib/, I've re-implemented it since Instr
is very simple (basically just string::find()) and we don't need some of the
complexity in Presto's implementation for handling a third argument.

Reviewed By: pedroerp

Differential Revision: D30355883

fbshipit-source-id: 3bc9880e0afbee8ad2e08b05a2e100629cd24e08