Increase compile, link, and test parallelism on CircleCI (#115)
Summary:
This PR doubles the container size and sets all parallelism flags (compile, high-mem compile, link, and test) to match the number of CPUs available.

I did not do this before because a lot of the build latency was coming from building duckdb, and increasing parallelism wouldn't have helped. This is no longer seems to be true.

Pull Request resolved: https://github.com/facebookincubator/velox/pull/115

Reviewed By: mbasmanova

Differential Revision: D30571063

Pulled By: funrollloops

fbshipit-source-id: ede5f9a9e968276a99778717e25df742fd6c1f3e