Properly use SelectivityVector in aggregations. (#114)
Summary:
Pull Request resolved: https://github.com/facebookincubator/velox/pull/114

SelectivityVector in aggregations was only properly used in update() methods, but not updateSingleGroup() where it was assumed to be 'allRows'.
With the coming implementation of Aggregation Mask (when aggregations could have a boolean column as a mask to indicate what rows to aggregate), we need to assume that SelectivityVector might not have all rows selected.
This diff is to make that change.

Reviewed By: tanjialiang

Differential Revision: D30569720

fbshipit-source-id: f8e2f4c4f8dfb3bc6fd36c06d25bdc8a442980a2