Adding support for aggregation masks. (#131)
Summary:
Pull Request resolved: https://github.com/facebookincubator/velox/pull/131

Pull Request resolved: https://github.com/facebookexternal/presto_cpp/pull/489

- Supporting aggregation masks.

Aggregation mask is a reference to one of the boolean projections which contains true for all rows that should be aggregated and false for rows that should not be.
One of the examples is this query:
  SELECT count(1), sum(IF(linenumber = 7, partkey, null)) FROM lineitem
It will have two projections for aggregation operator:
  'equal' = (linenumber = 7)
  'partkey' = partkey
'partkey' is used as a channel for 'sum' aggregation, while 'equal' as a mask.
Aggregation 'count' does not need any channels and just counts all rows.

- Fixing the aggregation pushdown with incomplete selectivity vector.

- Disabling aggregation pushdown for multi-group-by when we use aggregation masks.

- Adding ability to specify aggregation masks in PlanBuilder.

Reviewed By: mbasmanova

Differential Revision: D30624680

fbshipit-source-id: f9ac613c7699332bca2f3b0bc89f8394a9c956d7