Store std::exception_ptr on EvalCtx.errors (#157)
Summary:
Pull Request resolved: https://github.com/facebookincubator/velox/pull/157

EvalCtx.errors used to store only the exception strings found when
evaluation expressions, and hence, if the exception had to be re-thrown, the
original exception type would be lost. Changing the vector to capture the whole
exception (via std::exception_ptr) instead.
This uses an opaque type flat vector to store the exception objects.

Reviewed By: mbasmanova

Differential Revision: D30709206

fbshipit-source-id: eaf73630b880703ca48ac3aad52f5e14d328779a