Fix Rowgroup String Dictionary Read (#168)
Summary:
When a string dictionary column has an inDict stream, a false in that
stream offsets the dictionary index to refer to the rowgroup
dictionary. This offset was applied to the wrong element of the result
set if the RLE group was not the first in the read.

Pull Request resolved: https://github.com/facebookincubator/velox/pull/168

Reviewed By: spershin

Differential Revision: D30794871

Pulled By: oerling

fbshipit-source-id: ce0a6a3d41f4c0ff3285858b174344460e65bd03