Handle fully filter pushdown for metadata query rewrite
Previously, the metadata query rewrite rule was applied after the filter
pushdown rule. If a filter on a non-partition column was fully pushed
down to the connector, the metadata query rewrite rule was not aware of
it and could incorrectly enable the rewrite.

This commit moves the metadata query rewrite rule before the filter
pushdown rule to avoid this issue.

Note that this move has a side effect. The connector rule prunes the
partition lists with the partition column filters. Since the metadata
query rewrite is now applied before that, the partition lists are not
pruned. But this doesn't affect the correctness as the filter node is
kept to prune the list at runtime.