Support for ascii encoding for a subset of rows. (#32)
Summary:
Currently for any String vector we store state on whether its fully ascii or not. When a vector is fully ascii it allows us to use the ascii fast path for any string computation. Most datasets are predominantly ascii , however presence of even one utf string will cause it to take utf path. We now maintain asciiness for the whole Vector but also maintain state on the rows we have processed.

Note:
 - We now also store asciiStateMap which is a selectivity vector of all rows we have processed so far; if we get a Selectivity vector for a row that we havent seen then we return nullopt else we return asciiness of whole vector.
 - We dont store the encoding mode state in the vector anymore.

TODO: This PR doesnt address case when a Vector is reused ; It might be possible then to have old / invalid ascii state. This will be addressed in a follow up PR.

Pull Request resolved: https://github.com/facebookincubator/velox/pull/32

Reviewed By: mbasmanova

Differential Revision: D30460561

Pulled By: kgpai

fbshipit-source-id: c1a6854bf4a0e49ccf1e48df5acfea75b90d3258