[presto-orc] Support Column Layout in HiveWriter
OrcWriter uses multiple stream to represent a column.
Default OrcWriter policy is to layout the streams in the
increasing order of the size. But this spreads the reader
IO as either all streams of a column are accessed together.
This change exposes the config in the OrcWriter to verify
the change.