Fix array_intersect and width_bucket for dictionary encoded ArrayVectors (#178)
Summary:
array_intersect logic used a generic function 'toElementRows' that projected the function rows to the elements vectors of the left or right side arrays. If the left or right side vector was an array encoded in the dictionary, then the toElementRows projection didn't take into account the dictionary encoding involved.

This change fixes the toElementRows logic to handle dictionary encoded ArrayVectors, and uses the new logic in the array_intersect and width_bucket functions.

Pull Request resolved: https://github.com/facebookincubator/velox/pull/178

Reviewed By: pedroerp

Differential Revision: D30859364

Pulled By: mbasmanova

fbshipit-source-id: f2fa0fc19ab3598051bd59946496526b992de225