Fix nullable simple function bug
Summary:
Nullable simple functions (when callNullable() is defined instead of
call()) didn't work as expected. The parameters are taken as a pointer, but they
are never actually set to nullptr, carrying garbage if the input contains
nulls.

The underlying issue was in evalCtx->mayHaveNulls(), which had a weird semantic. Changing the flag name/semantic to nullsPruned() to better reflect the behavior, fixing the condition in VectorAdapter, and removing unnecessary var setter.

Reviewed By: mbasmanova

Differential Revision: D30829425

fbshipit-source-id: f54782d2e11f80cd9f18dea6443702bda72e86e0