evalWithMemo() shouldn't bypass null optimizations (#235)
Summary:
Pull Request resolved: https://github.com/facebookincubator/velox/pull/235

For subexpressions where memoization is enabled, the eval follows the
evalWithMemo() path, and that ends up calling evalAll() directly, bypassing the
null optimizations in evalWithNull().
.
Results are mostly the same, except for the case where the expression
deterministically throws, and the null optimization would make the eval skip
the processing of this expression altogether.

Reviewed By: mbasmanova

Differential Revision: D31008389

fbshipit-source-id: 565ac60601b8326a8d044559a86d098665aae39f