Fix flaky TestDistributedClusterStatsResource and TestQueues
TestDistributedClusterStatsResource's tests are flaky since we were not relying on waiting for dependent heartbeats to be completed. As part of this commit, we are have added required pre conditions before asserting. The timeout of one of the tests in TestQueues was recently reduced from 240_000 to 60_000 and it has become flaky as well. We are increasing the timeout to 120_000 since the test is creating multiple queries and waiting for different status and it may take more time sometimes.