Fix Proxy call from coordinator to resource manager
When client makes http request to ResourceGroupInfoStateResource,
coordinator makes http call to resource manager to fetch cluster level
resource group info. This causes failure in scenario, when internal https
communication is enabled. As part of the fix, we are sending the right scheme
in the request to avoid the failure