Use std::copy instead of memcpy (#250)
Summary:
Pull Request resolved: https://github.com/facebookincubator/velox/pull/250

I'm debugging a crash in DecodedVector where indices_ is unexpectedly null.  In
mode/dev, the program is killed before I get to the relevant crash if the src
parameter is null even if the size is zero. std::copy doesn't do this.

Performance of std::copy should be identical to memcpy.

Reviewed By: pedroerp

Differential Revision: D30978509

fbshipit-source-id: f1684f14b120d490fe427c1c09bbadb510dbba17