Fix Resource Group Distributed Queuing
When a resource group has multiple subGroup and first subGourp has no query to run,
internalRefreshStats method reset the parent subGroup isDirty flag as part of subGroup.updateEligiblity.
This leads the next subGroup updateEligiblity not called as we check for the parent group isDirty flag.

In this PR we are fixing that by checking the subGroup.isDirty flag and also reseting the flag once we
process all dirtySubGroup for the given group.