Fix coordinator - resource manager proxy communication
When https communication is enabled internally between coordinator - rm,
all http requests to coordinator fails when it needs to reach out to rm for info.
Reason: Coordinator uses http schema from the request and make a call to resource mamanger on secure port.
We are fixing the behavior and always use the schema supported by resource manager.