Buffer and spill only required channels for distinct aggregates
Distinct aggregate spilling logic buffers the input pages and spill them
when memoryRevoke is called. Currently, entire input page is buffered and
spilling. The aggregate is not going to comsume all the data in the page.
We can buffer and spill only the required channels from the input page.
This reduces the memory required for buffering data and also reduces spilled
bytes.