Dedup input data before buffering in distinct aggregates
SpillableFinalOnlyGroupedAccumulator buffers input pages in `rawInputs` and when
spill is triggered, these pages are spilled to disk. Currently, entire page is
buffered and spilled. For distinct aggregates, we can dedup rows within the same
groupId before buffering. This reduces memory used for buffering and also amount of
data to be spilled, if spill is triggered.