Track size of row in RowContainer (#312)
Summary:
Adds a uint32_t to the row for tracking width of out of line variable
length data. This is only added if there are variable lenght fields or
accumulators.

This allows controlling the output batch size for hash join or group
by and for getting predictable size chunks of data for spilling.

Follow-up PRs will add setting this in hash build, group by key insert
and when growing variable width accumulators.

Pull Request resolved: https://github.com/facebookincubator/velox/pull/312

Reviewed By: mbasmanova

Differential Revision: D31268602

Pulled By: oerling

fbshipit-source-id: 81c2da48b47cffc8e48b58c23b259bd0923af021