Use AsyncDataCache and read-ahead with Hive (#300)
Summary:
Uses AsyncDataCache and ScanTracker with Hive if the MappedMemory
given to the reader is a AsyncDataCache. If this is a regular
MappedMemory, uses BufferedInput with the given DataCache, if any.

Sets most tests to use AsyncDataCache as MappedMemory also if no
affinity.

Tests TableScan with both AsyncDataCache and regular MappedMemory +
SimpleLruDataCache.

Pull Request resolved: https://github.com/facebookincubator/velox/pull/300

Reviewed By: mbasmanova

Differential Revision: D31270284

Pulled By: oerling

fbshipit-source-id: f2958187a99eaf9013380035309ada979c992937