Validate return type on simple functions (#342)
Summary:
Pull Request resolved: https://github.com/facebookincubator/velox/pull/342

Expr execution used to crash if the type specified by the user didn't
match the expected function signature. Adding a check so that we throw a nice
exception. Note that this only works for simple functions; vectorized functions
will need a more complicated fix since we need to try to bind the available signatures.

Reviewed By: wenleix, OswinC, mbasmanova

Differential Revision: D31257222

fbshipit-source-id: a14fd52953d543cd199de42399c6c01d014d5403