Implement FixedSizeArray vector. (#367)
Summary:
Pull Request resolved: https://github.com/facebookincubator/velox/pull/367

Implement Vector operations for FixedSizeArray.  A FixedSizeArray should work
whereever an Array does.  For efficiency, we implement the fixed size variant
in ArrayVector rather than have all functions virtual and use a subclass.

Currently Presto page serializer uses a "sparse" format to represent null
entries, where they are not allocated space in the vector.  This is
incompatible with the Arrow format for Fixed-Size lists, see:
https://arrow.apache.org/docs/format/Columnar.html#fixed-size-list-layout

In Arrow, space needs to be allocated for null entries (non-sparse).  Since we
would like to keep all types of data readable with Presto/Spark and have a
fixed length array look like a variable length one, for now we continue to
leave out the nulls from the serialization (sparse encoding).

We can extend the Presto pager serializer in a seperate diff.  In the meantime,
exporting nullable fixed size arrays to Arrow will require a copy to embed
space for the null entries.  Once Presto page serializer supports fixed size
arrays natively, this becomes zero-copy again.  Note, this only apples to
nullable fixed size arrays, non-nullable ones are already Arrow compatbile.

As a further optimization once the page serialier is updated, we can in
principle save some memory by not allocating the backing arrays for rawOffsets_
and rawSizes_, but directly calculate them as width * index.  We could do this
today for non-nullable, but it is unclear if it is worth the complexity, so we
leave this for now.

Reviewed By: pedroerp

Differential Revision: D31366497

fbshipit-source-id: 26751694e558d05c37476914fa0a01fb53ae0a4b