Remove minAverageValueSize field from ColumnStatistics
ColumnStatistics minAverageValueSizeInBytes can be computed when
required. This is only used in readStripe to estimate the row size.
This is used only once, so no caching is required.

This change will modify the behavior of minAverageValueSizeInBytes
for String and Binary column types. For these types  sizes are computed
per row group (which could truncate) and they are summed and then
divided again. With this change they are accumulated and then divied
in the end, so they will be slightly larger.

Consider the String statistics with 2 row groups (ignore overhead bytes)
First row group: 5 values with 19 bytes
Second row group : 1 values with 11 bytes.

Previously it would compute
First row group : 3 bytes
second row group: 11 bytes

After merge it would produce: (5 * 3) + (1 * 11) = 26 bytes.
Merged row group stats : 26/6 = 4 bytes.

With the new code it would produce: (19 + 11)/6 = 5 bytes.

This is a very minor behavior change and should not impact anything.