Add timing of IO (data read/load). (#427)
Summary:
Pull Request resolved: https://github.com/facebookincubator/velox/pull/427

We use runtime stats to publish two new counters for operators: "dataSourceWallNanos" and "dataSourceLazyWallNanos".
The first one is for the direct load in the TableScan operator.
The second one is for the loads happening elsewhere through Lazy Vectors.
We use thread local and a tiny interface to implement the stats collection.
In case of significant overhead (I doubt it) we can easily disable the methods from using thread local and doing anything, using a define or altogether.
The counters should allow us to see how much time we spend on waiting on sync IO during queries.

Reviewed By: mbasmanova

Differential Revision: D31615769

fbshipit-source-id: eee5c23d0e6d2afc775e170e9acec00d7b4f4c69