Refactor Hive Parquet dereference pushdown optimizer rule
Currently the rule is focussed on Hive Parquet tables. It doesn't work
in any other connector other than Hive. Parquet is a common module between
Hive and Delta connectors. Most of the code in the Hive Parquet dereference
pushdown rule is common between Delta and Hive connectors.

Refactor the rule and create a pluggable points so that Hive and other
connectors that work with Parquet data (such as Delta) can extend it and
avoid rewriting the same code.