Reduce Long dictionary memory consumption
LongDictionary had the following data structures.
 Values -> Array of unique values (long [])
 ValuesToIndexMap -> Stores value to index in Values (Map<Long,Int>)

This consumed 2.5X space for long data type and about 5X for integer
data types. This change removes the ValuesToIndexMap and replaces it
with a open hashset. -1 represents an empty slot and other value
represents the index into the Values array. After this change the
memory consumption of long will be 1.5X and integers will be 3X.