fix applyFunctionWithPeeling with nulls (#528)
Summary:
applyFunctionWithPeeling uses rows instead of applyRows to generate newRows.

This can break with a SIGSEGV when the function is applied with newRows as it may incorrectly indicate certain values are non-null, when they are actually null leading to accessing data in surprising memory locations.  Particularly when the function uses the default null behavior and so doesn't check for nulls.

Pull Request resolved: https://github.com/facebookincubator/velox/pull/528

Reviewed By: mbasmanova

Differential Revision: D32014279

Pulled By: kevinwilfong

fbshipit-source-id: 5fb278deeabb1afc4f2565251bde8f6c0e7082f5