Skip query failure with OOM if memory revoke is pending
Current revoking logic gets the total query memory usage and if it is
greater than (maxTotalMemory * spark_revoking_threshold), then it will
trigger a revoke which runs in a separate thread. When the thread
executes, it will mark the operator that is using revocable memory
to perform a revoke. It does not trigger a revoke immediately in the
operator. When the operator is invoked again by the driver, the actual
revoke will be triggered and the operator will perform a spill. This
entire operation is async in nature.

Due to this behavior, its possible for a task to continue reserving
more memory and OOM before revoke is actually triggered. In classic,
we have a big revocable memory pool and tasks can expand in terms of
memory before revoke is triggered. In presto-on-spark, this buffer is
really small and thats why there is a higher chance for queries to
OOM before revoke is triggered.

The ideal fix would be to make the revoking logic synchronous but it
would be lot of effort to get it done and ensure that it does not impact
presto classic logic. Instead, we can try to delay the failure if a
revoke request submission is in progress or a submitted revoke is pending
as revoke will trigger soon and revocable memory will be released.