Optimize DictionaryCompressionOptimizer conversion code
DictionaryCompressionOptimizer uses set for allWriters and
directConversionCandidate. Both of them can be replaced with the list.
Iterating all elements in list is much faster than a set.

tryConvertToDirect removed dictionary columns from inside the method.
Now the reponsbility is shifted to the caller.
convertLowCompressionStreams uses an iterator and removes the iterator.
Previously this required a copy as the removal while iterating  produced
concurrentModificationException.