Fix variantArrayToVector for large strings (#547)
Summary:
FlatVector<StringView::set() copies the string into internal string buffers,
hence, this method is safe and the caller doesn't need to worry about the
lifetime of the buffer referenced by the StringView. variantArrayToVector was
using this method as if it worked like setNoCopy which doesn't copy the string.
variantArrayToVector was overwriting the internal string buffers making
StringView entries in the vector point to memory that's no longer owned by
anyone.

Pull Request resolved: https://github.com/facebookincubator/velox/pull/547

Reviewed By: funrollloops

Differential Revision: D32123092

Pulled By: mbasmanova

fbshipit-source-id: 9ec663df26d9d7d37fdbcb25dca47a06f9fd4c38