Reject list for map key filtering (#503)
Summary:
Pull Request resolved: https://github.com/facebookincubator/velox/pull/503

# Problem
Currently we can read a flatmap with a key filter. This filter only works as an allow list. You have to specify what you want to read from a map column.

I need a feature to expand this to support a reject list. This is reasonable but I need some input about how to pass the information.

# Solution
Currently we use `ColumnSelector` to pass an expression, as a json array.

```
columnName#[1,2,3,4]
```
In this diff, I'm proposing this:
```
columnName#["!1","!2"]
```
To read `columnName` all keys but `1`, and `2`.
I'm not sure if this is a right choice, and not sure if we can do this for string key type. I only used `!` for integer keys when they came as string type for now.

Reviewed By: harsharastogi

Differential Revision: D31912159

fbshipit-source-id: 1a84fa3a427dcceb7d5e59a3f012074fb9bcd35a