Tokenization logic to support Joda datetime (#555)
Summary:
Pull Request resolved: https://github.com/facebookincubator/velox/pull/555

Tokenization logic to support Joda datetime, in order to eventually
support Presto's parse_datetime() with the same semantic as in Java.

This diff only contains the tokenization logic; the next one in the stack
will add the parsing logic based on the tokens identified in this diff, then
the logic to convert the parsed input back to a timestamp.

The tokenization logic follows the definition in:
> http://joda-time.sourceforge.net/apidocs/org/joda/time/format/DateTimeFormat.html

Reviewed By: mbasmanova

Differential Revision: D32013564

fbshipit-source-id: 1e1fbc007dbdd9df21569e449652ba6d5012ec45