Fix IN predicate for Spark (#565)
Summary:
Pull Request resolved: https://github.com/facebookincubator/velox/pull/565

Vector function that has a single argument is guaranteed to receive its input as
flat. Vector function that has multiple arguments all but one of which are
constant is guaranteed to receive its not-necessarily-constant argument as flat
or constant. However, in this context "constant" doesn't mean vector of
constant encoding. Instead, it means that the vector has same value for all
rows in scope, e.g. BaseVector::isConstant(rows) is true.

Reviewed By: funrollloops

Differential Revision: D32197009

fbshipit-source-id: 1f2d72b2dfe89f7cffcd00fc011d132024b9d43e