Add rpad and lpad Presto functions (#556)
Summary:
Pull Request resolved: https://github.com/facebookexternal/presto_cpp/pull/556

Add functions to reproduce the behavior of Presto's

lpad(string, size, padstring) → varchar#
Left pads string to size characters with padstring. If size is less than the length of string, the result is truncated to size characters. size must not be negative and padstring must be non-empty.

rpad(string, size, padstring) → varchar#
Right pads string to size characters with padstring. If size is less than the length of string, the result is truncated to size characters. size must not be negative and padstring must be non-empty.

Pull Request resolved: https://github.com/facebookincubator/velox/pull/506

Reviewed By: mbasmanova

Differential Revision: D31938358

Pulled By: kevinwilfong

fbshipit-source-id: 7aab1ea1c1b4b8f063f04e7c3c5226904b0fa815