Add input_mtimes and output_mtimes for DEPA hive table
To support this, we mainly made following changes

For DDL queries, we add MetastoreOperationStats to take the lastDataCommitTime from
metastore, and return it to SemiTransactionalHiveMetastore, where they are compiled into
ConnectorCommitResult and returned to TransactionManager, and further to PrismEventListener.

For SELECT queries, the compiler gets the partition names of the input table, and then we fetch
and attach the lastAccessTime from metastore for them.

For CTAS queries, it should be executed as SELECT queries for input tables and DDL queries for
output tables.

TESTS:
1. Ran PrismQueryRunner in local to check the logging.
2. Deployed to testing cluster to run various queries, and checked the output of DEPA table.