Use diamond operator where possible
The diamond operator, introduced in Java 7, adds type inference and
reduces verbosity in assignments when using generics. The vast majority
of this codebase uses the diamond operator where possible, but a few
legacy usages of the earlier style remain. This modernizes the code by
adopting the diamond operator where possible, improving legibility
through the use of a consistent programming paradigm.