Add delay for checking postconditions of DriverTest (#586)
Summary:
Some postconditions of completing or cancelling a query are realized
slightly after the last thread acknowledges the stop or a cursor is
marked as being at end. Specially the latter case is not practical to
be fused in the same critical section that marks the result set as
complete. We add a macro to test the postcondition with a delay so as
to reduce test flakiness.

Pull Request resolved: https://github.com/facebookincubator/velox/pull/586

Reviewed By: mbasmanova

Differential Revision: D32256530

Pulled By: oerling

fbshipit-source-id: 05b082f321570786047a45358402c0d8525259dd