Install brew package with a different version (#621)
Summary:
Add support to install a brew package with a different version in setup-macos.sh
This useful when there is some problem with the latest version from brew so you wan to pin the build to a certain old version.
After this PR, you just add something like "package:version" to the "end" of MACOS_DEPS list, e.g.:
MACOS_DEPS="... antlr4-cpp-runtime:4.9.2"
Note that the ':' is used to separate package name and version. This is different from '@' where '@' is actually part of the formula name in brew.

This the article I read as a reference on how to install a brew package with a specific older version:
https://cmichel.io/how-to-install-an-old-package-version-with-brew/

Pull Request resolved: https://github.com/facebookincubator/velox/pull/621

Reviewed By: funrollloops

Differential Revision: D32466921

Pulled By: mshang816

fbshipit-source-id: f8e2bedb28d154192d02951df0865e9a15816ea7