Refactor and add option to build Velox test utilities (#596)
Summary:
Velox test utilities are helpful to test Velox integrations with other tools such as presto_cpp.

The scope of this change is to allow building the test utilities as a standalone library.
As part of this work, some helper classes in `velox/exec/tests` have been moved to the `velox/exec/tests/utils` folder.

Pull Request resolved: https://github.com/facebookincubator/velox/pull/596

Reviewed By: mbasmanova

Differential Revision: D32368802

Pulled By: kgpai

fbshipit-source-id: eed0ab2ed1a1212e85ee04694c7c0c4c9a8924ba