Add AVX2 implementation of BigintValuesUsingHashTable (#614)
Summary:
Does range checks and index calculation and load of first hit
candidate 4 values at a time. If the initial load gets a hit or an
empty marker in every lane, the IN is resolved. If there are lanes
that were neither a match of the value nor an empty marker, we load
the next 4 values and check if either a hit or empty marker is found
in them. We do so until either a hit or empty marker is found. We pad
the 4 values above the hash table's last value to be equal to the last
value so that we do not have to wrap around on the first load after a
collision but only after that.

Evaluating the scalar IN filter is up to 10% of e.g. Baymax CPU and is
generally a common operation.

Adds a benchmark to compare with the scalar path:

../../velox/type/tests/FilterBenchmark.cpp      relative  time/iter  iters/s
============================================================================
scalarDense                                                  2.72ms   367.80
simdDense                                        301.84%   900.75us    1.11K
scalarSparse                                                 1.83ms   546.22
simdSparse                                       208.73%   877.10us    1.14K

Pull Request resolved: https://github.com/facebookincubator/velox/pull/614

Reviewed By: pedroerp

Differential Revision: D32430727

Pulled By: oerling

fbshipit-source-id: 911c29a3383ed233271859d0426361875df1f4a9