Fix ASAN failure in compacting nulls in SelectiveColumnReader (#629)
Summary:
If a filtered column is followed by another filtered column, the
non-last filtered columns will be compacted to align with the last
filtered column's passing rows. In doing this, nulls must be compacted
only if there actualxly are nulls. There can be nulls because some
filters will pass nulls. The error occurs when there were nulls in a
previous batch and there are no nulls in a subsequent batch and the
nulls uffer is too small, leading to a rare access past the end of the
resultNulls_ buffer.

Pull Request resolved: https://github.com/facebookincubator/velox/pull/629

Reviewed By: zzhao0

Differential Revision: D32494125

Pulled By: oerling

fbshipit-source-id: a0d56cfa9ae69ac270f1681412627dc36c8b9f05