Implement weighted split scheduling in presto-main and presto-spi
This change adds the notion of a SplitWeight to the presto-spi, a
concept which allows connectors to indicate the amount of work
associated with a given split, relative to some "standard" split.

Connectors can choose to assign that concept of a weight based on
whatever heuristics might be appropriate within the context of that
connector, such as size in bytes or some other metric that might
be known about the split at the time that it is generated. The
calculation need not be fully precise, but care should be taken
when implementing weight calculations in connectors to avoid using
weights that are very small or very large.

On the presto-main side, split weights are used to inform how many
splits are allowed to be running on a given node or queued on a
given task (still configured by: node-scheduler.max-splits-per-node
and node-scheduler.max-pending-splits-per-task respectively). Those
values are now interpreted to be relative to one "standard" split
weight. When all splits are assigned the standard weight, the
scheduler will behave the same way as before. However, when splits
are assigned weights that are smaller, the scheduler will allow
more of them to be assigned or queued to tasks. In effect, this
allows the coordinator to assign enough splits to workers for them
to stay busy between batches of split assignments when the amount
of time that workers take to complete individual splits is too
short.

In order to control the maximum number of splits that might be
delivered in a single task update request, the existing config
parameter node-scheduler.max-unacknowledged-splits-per-task still
controls the absolute count of splits that the coordinator will
allow to be assigned to a given task that have not yet been
acknowledged by the worker as having been received. This can be
especially important if splits themselves have a large serialized
JSON representation, in which case sending a large number of small
splits (by weight) could create huge task update requests to be
sent.