Implement size based split weights for presto-hive
Implements split weighting for HiveSplit instances based on their
size. This behavior is enabled by default and can be disabled by
setting the session property `size_based_split_weights_enabled=false`
or the hive configuration property `hive.size-based-split-weights-enabled=false`.

When enabled, splits are assigned their weight based on their size
in bytes relative to `hive.max-split-size`. Splits that are 1/2 of
the max split size will be weighted as 1/2 of the standard split weight.

Splits that are larger in size than the hive target split size
(eg: unsplittable files) are still assigned the standard split weight
to avoid scheduler interactions that might be harmful when extremely
large weights are calculated. This is a conservative decision designed
to prevent existing workloads from regressing, but might be work
revisiting in the future.

Finally, when size based split weights are enabled, no split will be
assigned a weight smaller than the value set by the
`minimum_assigned_split_weight` session property or the
`hive.minimum-assigned-split-weight` configuration property which
defaults to 0.05 (proportional to 1.0, the weight of a standard split).
This provides a mechanism to control how aggressively the scheduler
will respond to the presence of small files. With the default
configuration, files smaller than 5% of the target split size will
still be assigned 5% of the standard split weight, allowing a
maximum of 20x more splits to be running or queued.