Fix AbstractArrayBlock#copyPositions with zero elements
AbstractArrayBlock#copyPositions could potentially fail when no
underlying element positions were selected (either because selected
array positions were all null, all empty arrays, or copyPositions
was passed a length argument of 0). When that occurred, the raw
element block would receive a copyRegion call of length 0, but with
the starting offset of the ArrayBlock offsets base which might not
be a valid position within the elements block. The fix is to call
copyRegion from position 0 of length 0, which should be valid in
all scenarios with the underlying block.

Also adds a test to verify the fix.