Fix reading fixed_len_byte_array of length 9-16 in parquet
Cherry-pick of https://github.com/trinodb/trino/pull/9971

If the extra bytes are all zeros or all -1 for negative values,
then the remaining bytes can be decoded as a short decimal.

Currently incorrect results are returned when a negative value is
stored in fixed_len_byte_array of size 9-16 and read as a short decimal.

Co-authored-by: Raunaq Morarka <raunaqmorarka@gmail.com>