Optimize Single Type Writer Constructors
SingleArrayBlockWriter, SingleMapBlockWriter and SingleRowBlockWriter in
their constructor, computes the initial size. Late if getSizeInBytes is
called, it can compute the size of the object by getting the current
size and subtracting the initial size. But single type writers are short
lived objects and never used for getSizeInBytes. BlockBuilder inherits
Block interface, so the calculation should be deferred until called.
This change defers the getSizeInBytes calculation, until required.