Fix unneeded signed/unsigned conversion
Summary:
Start working down -Wsign-compare warnings in hopes of being able to use -Wextra in the future.

The findNextReplaced() lambda just wraps findNthInstanceByteIntex, which returns a signed long. However the return value is stored in a size_t, which is unsigned.

Use "auto" for the type instead, so that it's stored as a signed. That allows the later bounds check (== -1) to make semantic sense without having to reason about signed/unsigned comparison.

The old behavior here is well defined, so this change is a no-op

Differential Revision: D31907193

fbshipit-source-id: 51544a48098dbfa06121e688c08098b18a4b3c5c