Do not use MapVector::canonicalize outside of construction (#655)
Summary:
Comparing or serializing data to RowContainer must not mutate the
source data. Maps must be compared and serialized in key order but
canonicalize cannot be used for this because this will lead to flaky
crashes if different threads handle one MapVector, which can happen
for example with local exchange.

Use a commutative hash mix for maps since the key order is uncertain.

Covered by exec/tests/RowContainerTest and vector/tests/VectorTest.cpp.

Pull Request resolved: https://github.com/facebookincubator/velox/pull/655

Reviewed By: mbasmanova

Differential Revision: D32615991

Pulled By: oerling

fbshipit-source-id: 4322330846b8957910d93dff22816145373e286b