Reduce dictionary row group size
Dictionary row group consumes large amount of memory when the indexes
repeat. Consider a single column of String where the number of unique
values of string is 128. getBufferedBytes will return  number of
rows + size of dictionary. But the actual memory consumption would be
4 * number of rows + size of dictionary. On some tables with 100's of
columns, this is 100's of MB of difference and causing OOM issues.

Java generics does not work on primitive array so the code is
duplicated for byte, short and int. This introduces lot of code
duplication, but reduces the memory required significantly.