Rename FunctionRegistry.h to ScalarFunctionRegistry.h (#748)
Summary:
Pull Request resolved: https://github.com/facebookincubator/velox/pull/748

High Level Goal:
I want to change ScalarFunctions and AdaptedVectorFunctions to do function call lookups based on FunctionSignatures.

This is primarily because the current solution of doing exact match lookups against the argument types is overly restrictive and makes adding features like VariadicArgs difficult.  SignatureBinder is much more flexible and has the flexibility to allow for more features in the future.

With VectorFunctions already using FunctionSignatures, and AggregateFunctions in the process of migrating, this also represents a step towards consistency in behavior.

This change:
I plan to pull out the common logic from FunctionRegistry.h and VectorFunctionBindings.h into a new file, along with the common code to create the new registration system.  I'd like to call this new class/file FunctionRegistry.

The existing FunctionRegistry.h will only contain the declaration of ScalarFunctionRegistry, the ScalarFunctions() function, and a stripped down version of registerFunction.  I thought it would be better named ScalarFunctionRegistry.h, to match the existing VectorFunctionRegistry.h for AdaptedVectorFunctions

This diff just renames the file to keep the change from cluttering downstream diffs.

Reviewed By: kgpai

Differential Revision: D33078007

fbshipit-source-id: 636430cec803bef3b9966907044ee157132eb52e