Fix DriverTest end of result set (#740)
Summary:
DriverTest.pause and possibly other tests can be flaky so that the end
state of the task may not be kFinished and the final Driver count may
not be 0. The flakiness happens because destruction of TaskCursor
calls requestTerminate on the Task after the end of the result set. At
this time, there can be threads inside the Task. If so, these take an
error path to exit and may leave the state to be kCancelled or
kFailed.  The fix is to request terminate on the Task at TaskCursor
destruction only if there is unread output. If all output is consumed,
the Task will go off thread and transit to kFinished in a short time.

To check that the end state is correct, we take a stateChangeFuture()
and wait for a non-running state, which is the official API for
listening to state changes.

Pull Request resolved: https://github.com/facebookincubator/velox/pull/740

Reviewed By: mbasmanova

Differential Revision: D33063960

Pulled By: oerling

fbshipit-source-id: 6e41e45f05e7e27f505bbf1d888868a9c575fca5