Fix output layout handling in LocalPartition operator (#791)
Summary:
Output layout (the order of columns in the output) applies to results, but LocalPartition was applying it to input. In case when partitioning key had a different position in the output, hashing was applied to the wrong column potentially of the wrong type leading to a crash in DecodedVector::valueAt<T>().

Pull Request resolved: https://github.com/facebookincubator/velox/pull/791

Reviewed By: spershin

Differential Revision: D33255199

Pulled By: mbasmanova

fbshipit-source-id: 2136621836b638063621e2d41757de7ecc1cfb20