Expose nested BlockBuilders for nested types
https://github.com/prestodb/presto/pull/16965 exposed keyBlockBuilder
and valueBlockBuilder. That PR worked when Map type is one of the top
level columns. But if the Map type is nested inside Array, calling
beginBlockEntry on array always returned SingleMapBlockWriter.
This change exposes the nested block builder from all the nested types.
There are tests written to show how to use them.

This change avoids creation of temporary objects (Single*Writer) for
all nested types. But writer that requires to use this should have
the knowledge of the type and should cast objects to the required type.