Add TPCDS schema to HiveQueryRunner
This commit enables HiveQueryRunner with TPCDS schema.

In addition, we have to make TestHiveLogicalPlanner single-threaded to
resolve the resource contention issue when both UT
testBaseToViewConversionWithDerivedFields and
UT testBaseToViewConversionWithGroupBy run in parallels. There could be
an issue on local resource conflict without serialization when running
UTs in parallels. Changing the class to single-threaded will avoid
such problem.