Add SsdFile - A Shard of a SSD cache for AsyncDataCache (#780)
Summary:
Entries from AsyncDataCache can be written to and read from an
SsdFile. When writing, the file allocates space for the cache entries,
possibly growing the file or evicting older entries.  When reading,
the caller obtains SsdPins for the entries to be read. The entries are
guaranteed to stay live for the duration of their pin. The entries can
then be read with a coalesced IO.

The SsdFile is composed of regions. Each region has entries and
entries do not span regions. The region is the unit of eviction and
pinning and access tracking. When the cache is read from, region read
counts are updated and older counts are decayed. When needing to write
to a full SsdFile, we take the lowest score unpinned region and remove
the entries in this region from the key to entry map and rewrite the
region.

Adds a utility function readPins for coalescing read of multiple pins
from a file or other container supporting ReadFile::preadv or similar.

Pull Request resolved: https://github.com/facebookincubator/velox/pull/780

Reviewed By: mbasmanova

Differential Revision: D33285585

Pulled By: oerling

fbshipit-source-id: 6c297391ed5dbfcf2c18fbc19ea6a4f556651e41