Change LEFT join to preserve the order of input rows in the output (#834)
Summary:
Presto query planner relies on left join to preserve the order of probe-side rows, e.g. if probe-side rows are grouped or sorted on column A, the output will be also be grouped or sorted on column A. This is used when propagating "grouped" property of the probe-side columns for the purpose of enabling streaming aggregation.

Pull Request resolved: https://github.com/facebookincubator/velox/pull/834

Reviewed By: oerling

Differential Revision: D33420124

Pulled By: mbasmanova

fbshipit-source-id: 5ed0eff3888535b63d15a1412b6f8e4f225d36e2