Add option to resize without initializing nulls. (#852)
Summary:
Resizing a vector sets all the nulls to not null in the new added rows.
However, in cases such as in vector function adapter that writes the nullity twice.

Setting nulls cant be avoided because sometimes the vector is pre-allocated with undefined null state.
Also resize(0) and resize(X) does not work because sometimes we need to keep previously written data to be valid.

This diff adds a flag to BaseVector::resize() that can tell the function not to set nulls.
Also change the usages in the vector adapter to resize with out setting nulls.

It is expected that scalar functions producing complex type results would resize without setting nulls.
A series of follow-up diffs is needed to revisit the usage of resize() in existing function and change as appropriate.

Pull Request resolved: https://github.com/facebookincubator/velox/pull/852

Reviewed By: mbasmanova

Differential Revision: D33464472

Pulled By: laithsakka

fbshipit-source-id: cb0aa2810fae991a647791a8af184eee1c7ae1a2